1. Refactored code. 2. A bug regarding initializing esbuild mutiple times does not solve.
