+++ package.json
@@ -11,6 +11,8 @@
     "@testing-library/user-event": "^13.5.0",
     "@types/jscodeshift": "^0.11.5",
     "@types/prettier": "^2.7.1",
+    "@types/react": "^18.0.21",
+    "@types/react-dom": "^18.0.6",
     "@types/react-monaco-editor": "^0.16.0",
     "axios": "^1.1.3",
     "buffer": "^6.0.3",
